Added mechanism to specify dynamic workflow-specific page titles to job submission pages.

+++ src/main/java/edu/yu/einstein/wasp/load/WaspLoadService.java
@@ -82,6 +82,9 @@ public abstract class WaspLoadService {
 		m.put("area", area);
 		List<UiField> oldUiFields = uiFieldService.findByMap(m);
 		for (UiField uiField: oldUiFields) {
+			if (uiField.getName().indexOf("jobsubmit/")>-1) {
+				continue;//do NOT remove custom titles
+			}
 			uiFieldService.remove(uiField);
 			uiFieldService.flush(uiField);
 		}
